+++ packages/better-effect-kysely/src/index.ts
@@ -2,6 +2,8 @@ export { KyselyEffect } from './service'
 export { KyselyQueryError, KyselyTransactionError } from './errors'
 
 export type { KyselyServiceFactory } from './service'
+export type { KyselyExecutable } from './execute'
+export type { KyselyTakeFirstExecutable } from './execute-take-first'
 export type { KyselyExecutionOptions } from './options'
 export type { KyselyOperation } from './operation'
 export type { KyselyQueryOperation } from './errors'